ctf2/hospital/all_requirements.txt
Babel==1.3
BeautifulSoup==3.2.1
CairoSVG==1.0.9
CoAPthon==4.0.0
Coffin==2.0.1
Django==1.6.11
ESClient==0.5.8
Fabric==1.10.0
Flask==0.10.1
Flask-Admin==1.0.8
Flask-ESClient==0.1.1
Flask-Mail==0.9.1
Flask-MySQL==1.2
Flask-OpenID==1.2.4
Flask-RESTful==0.3.5
Flask-Script==2.0.5
Flask-WTF==0.10.2
Jinja2==2.7.3
M2CryptoWin32==0.21.1-3
Markdown==2.5.2
MarkupSafe==0.23
Mezzanine==3.1.10
Pillow==2.9.0
PyPDF2==1.25
PySimpleSOAP==1.16
Pygments==2.0.2
Pyphen==0.9.1
Sphinx==1.2.2
UNKNOWN==0.0.0
WMI==1.4.9
WTForms==2.0.1
WeasyPrint==0.23
Werkzeug==0.9.6
XlsxWriter==0.7.3
activedirectory==0.14
akismet==0.2.0
alabaster==0.7.3
amqp==1.4.6
aniso8601==1.1.0
anyjson==0.3.3
argparse==1.3.0
beautifulsoup4==4.1.3
billiard==3.3.0.20
bleach==1.4.1
blinker==1.3
cairocffi==0.6
cassandra-driver==2.6.0
celery==3.1.18
cffi==0.8.6
chardet==2.3.0
click==3.3
colorama==0.3.3
cryptography==0.8
cssselect==0.9.1
defusedxml==0.4.1
diff-match-patch==20121119
django-aggregate-if==0.5
django-ajax-chat==0.2
django-ajax-filtered-fields==0.5
django-ajax-selects==1.3.5
django-ajax-upload-widget==0.5.5
django-ajaximage==0.2.8
django-appconf==1.0.1
django-autofixture==0.9.2
django-batch-select==0.2.4
django-bootstrap3==4.10.0
django-cassandra-engine==0.4.0
django-celery==3.1.16
django-compressor==1.4
django-contrib-comments==1.6.1
django-crispy-forms==1.4.0
django-custom-user==0.5
django-encrypted-fields==1.1.1
django-extensions==1.5.0
django-filebrowser==3.5.6
django-fluent-comments==0.9.2
django-formset-js==0.4.3
django-grappelli==2.5.3
django-guardian==1.2.4
django-hstore==1.3
django-hvad==1.2.0
django-jinja==1.4.1
django-jquery-js==2.1.3
django-jsonfield==0.9.13
django-libs==1.66.1
django-mail-queue==2.2.2
django-mailchimp-v1.3==1.3.1
django-mailer==0.2a2
django-markdown==0.8.0
django-markdown-deux==1.0.5
django-model-utils==2.2
django-mongodb-engine==0.5.1
django-mutant==0.1.2
django-nested-inlines==0.1
django-otp==0.2.7
django-pagedown==0.1.0
django-pandas==0.2.2
django-picklefield==0.3.1
django-polymodels==1.2.2
django-queryinspect==0.0.4
django-redis==3.7.1
django-registration==0.8
django-registration-email==0.5.4
django-reversion==1.8.1
django-reversion-compare==0.3.5
django-simple-captcha==0.4.4
django-social-auth==0.7.25
django-social-media-links==0.2
django-social-share==0.3.0
django-sslify==0.2.7
django-tables2==0.15.0
django-taggit==0.12
django-tastypie==0.12.2
django-threadedcomments==0.9.0
django-tinymce==1.5.3
django-two-factor-auth==1.1.1
django-versionfield==0.3.2
django-wysiwyg==0.7.0
djangotoolbox==1.6.2
docutils==0.12
dodge==0.1.7
ecdsa==0.11
elasticsearch==1.2.0
elasticutils==0.10.2
enum34==1.0.4
eventlog==0.8.0
extras-mongoengine==0.1
factory-boy==2.5.1
fdfgen==0.11.0
filebrowser-safe==0.4.0
filemagic==1.6
flask-elasticsearch==0.1
flask-login==0.2.11
flask-mongoengine==0.7.1
flextls==0.3
future==0.9.0
futures==3.0.3
google-api-python-client==1.3.2
graphviz==0.3.4
grappelli-safe==0.4.1
html5lib==0.999999
httpagentparser==1.7.6
httplib2==0.9
ipaddress==1.0.6
ipdb==0.8
ipython==2.1.0
itsdangerous==0.24
jdcal==1.0
jsonfield==1.0.0
kombu==3.0.26
ldap3==0.9.7.11
logger==1.4
lxml==3.3.5
mailer==0.7
mammoth==0.3.8
markdown2==2.3.0
matplotlib==1.4.2
mechanize==0.2.5
memory-profiler==0.32
mezzanine-flexipage==0.2.0
mezzanine-pagedown==0.8
mongodbforms==0.3
mongoengine==0.8.7
mysql-connector-python==2.0.3
mysql-python==1.2.5
natsort==3.5.6
netifaces==0.10.4
numpy==1.9.1
oauth2==1.9.0.post1
oauth2client==1.4.6
oauthlib==1.0.3
openpyxl==2.1.5
openvas.omplib==0.1.0
pandas==0.16.1
paramiko==1.14.1
parsimonious==0.5
pdfkit==0.4.1
pisa==3.0.33
plotly==1.7.6
ply==3.4
preppy==2.3.2
psutil==2.1.3
psycopg2==2.6.1
py2docx==0.2.0
py2exe==0.6.9
py4j==0.8.2.1
pyOpenSSL==0.14
pyad==0.5.14
pyasn1==0.1.7
pyasn1-modules==0.0.5
pycparser==2.10
pycrypto==2.6.1
pydot==1.0.2
pyelasticsearch==0.7.1
pymongo==2.7.2
pyparsing==1.5.7
pyreadline==2.0
pyroute2==0.3.5
python-dateutil==2.4.2
python-docx==0.7.6
python-keyczar==0.71c
python-ldap==2.4.19
python-libnmap==0.6.2
python-linkedin==4.1
python-magic==0.4.6
python-mimeparse==0.1.4
python-openid==2.2.5
pytz==2015.4
pyusb==1.0.0rc1
pywin32==213
qrcode==4.0.4
redis==2.10.3
reportlab==3.1.8
requests==2.7.0
requests-oauthlib==0.5.0
rsa==3.1.4
schedule==0.3.0
selenium==2.44.0
shodan==1.3.1
simplejson==3.6.5
six==1.9.0
slimit==0.8.1
snowballstemmer==1.2.0
soap2py==1.16
sphinx-rtd-theme==0.1.7
sqlparse==0.1.13
stripe==1.32.2
svg2rlg==0.3
svn==0.3.25
termcolor==1.1.0
timer==0.1
tinycss==0.3
trml2pdf==0.2
tzlocal==1.0
udemy-dl==0.1.7
uritemplate==0.6
urllib3==1.9
virtualenv==13.1.2
wget==2.2
wkhtmltopdf==0.2
xhtml2pdf==0.0.6
xlrd==0.9.3
xlutils==1.7.1
xlwt==0.7.5
z3c.rml==2.9.2
zope.event==4.0.3
zope.interface==4.1.2
zope.schema==4.4.2
